restructuring code in preparation for architecture dev
